Extract _record_for from populate, use round trip in _apply_Update
This will ensure that we have exactly the same logic/behavior across the
board when turning records into content and prevent the :-( hack that was in
here before. It was missing the max(min, ttl) bit we throw everything else
through and this makes that consistent. Most importantly it'll prevent us
from having to fix bugs or make improvements in multiple code paths in the
future.