Adding Octodns provider class for easyDNS
This provider class for easydns.com adds support for basic dns records
through the easyDNS v3 API.  Support for dynamic and geo based dns
records is planned for a future update.

Sample configuration for the easyDNS provider are:

  easydns:
     class: octodns.provider.easydns.EasyDNSProvider
     token: <token>
     apikey: <key>

The token and key values are found on the easyDNS customer portal at:

https://cp.easydns.com/manage/security/api/production_info.php

Also, below are some optional configuration parameters which can be
added to override the class defaults.  By default the provider class
connects with the LIVE easyDNS API, if you wish to perform testing
with the easyDNS Sandbox API you can enable it by adding the following
configuration parameter:

     sandbox: True

Note, the API token and key are different for the sandbox than they
are for the production API, you can obtain sandbox credentials at:

https://cp.easydns.com/manage/security/api/sandbox_info.php

Lastly, if you have created Domain Portfolios through the easyDNS CP
you can configure which portfolio new domains will be added to by
supplying the portfolio option with the name of your portfolio.

     portfolio: <portfolio name>