Improve ALIAS, CNAME, DNAME & PTR record FQDN validation
Use fqdn package to help verify if the record value is really valid.

The original behavior will treat value like `_.` or `.` be a valid
record, which is strange, and the real world may not have those use
cases at all.

The RFC documents are pretty long, as I didn't read them all or enough
to tell should it be valid or not by the spec, so I opened issue #612 to
discuss this case and got a positive response from the main maintainer
to have the change.

Close #628