Improve checking and creating Azure DNS Zones
This change improves the process for checking for AzureDNS zones by using the known set and not relying upon custom error handling.

Since the provider already fetches the zones, octodns doesn't need to make a second call to check for the existence of the zone - _populate_zones already does that for us.