Change DNSimple's Provider to query zones instead domains
A quick summary of a problem with have with DNSimple provider:
Let's suppose that I have the following config:

zones:
  30.114.195.in-addr.arpa.:
    sources:
      - config
    targets:
      - dnsimple

Even if a customer has this Reverse zone configured in DNSimple, this
fails with:

400 Bad Request for url: https://api.sandbox.dnsimple.com/v2/x/domains

because it is trying to create a domain because the zone wasn't found.
octodns.provider.dnsimple.DnsimpleClientNotFound: Not found

This happens because the GET /domains endpoint at DNSimple does not
bring Reverse Zones. To make this work nice, we should use /zones/
instead making it return properly the reverse zones.