Prepare tests with failing "managed subzone" error
The is not a zone between delegated.subzone.unit.tests. and unit.tests.,
however we get a delegated subzone error.  This modifies the tests to
succeed with the added record, however the tests fail as it incorrectly
throws the managed subzone error.

Change the name of delegated.subzone, and the tests will pass cleanly.