Fix issue with using Templating processor on aliased zones
Switches the processor to use process_source_and_target_zones which
happens later, during planning changes and thus happens seperately for
the source and aliased zones. This leaves the templates in the copy of
desired that Manager uses to start the alias process. The updated copy
is in the plan so externally everything will still make sense and
behave as expected.