processors: Added a simple TTL clamping processor
This will come in handy for APIs not supporting TTL outside a certain range.
The Spaceship API, for example, only allows TTLs in the range of 5..3600 but rewriting a
whole zone for Spaceship only seems not as convenient as clamping the values as they flow
through OctoDNS.

The code is coming from Claude. My prompt was:

    Write a simple OctoDNS processor that clamps TTL values.

I followed up with:

    AttributeError: 'TtlClampProcessor' object has no attribute 'log'
    Where should self.log come from?

And finally:

    Hm. I suspect the code is not running, somehow. It doesn't seem to be getting control. I
    put a raise in the process_record method but it doesn't fail. And I don't see any clamped
    values nor log output.

So it took three attempts to make it produce something useful.